broke up collector
